src/main/java/org/jugistanbul/secondopinion/api/validator/MedicValidator.java
package org.jugistanbul.secondopinion.api.validator;

import org.jugistanbul.secondopinion.api.entity.Medic;
import org.springframework.stereotype.Component;
import org.springframework.validation.Errors;
import org.springframework.validation.ValidationUtils;
import org.springframework.validation.Validator;
@Component
public class MedicValidator implements Validator {
	@Override
	public boolean supports(Class<?> clazz) {
		return Medic.class.isAssignableFrom(clazz);
	}

	@Override
	public void validate(Object target, Errors errors) {
		ValidationUtils.rejectIfEmptyOrWhitespace(errors, "username", "error.username", "Username is required.");		
		ValidationUtils.rejectIfEmptyOrWhitespace(errors, "email", "error.email", "Email is required.");		
		ValidationUtils.rejectIfEmptyOrWhitespace(errors, "password", "error.password", "Password is required.");		
	}

	

}
